osm2railjson: move switch and buffer stop code
Move switch and buffer stop code to a separate file.
